codec change attempts
